add seed

+++ resources/coconut.rb
@@ -2,7 +2,7 @@
 class Coconut
   include Mongoid::Document
   field :name, type: String
-  field :mass, type: Integer
+  field :mass, type: Float
   field :juicy, type: Boolean
   field :keywords, type: Array
 end
